Lift authorities from measurement-map
The authorities that must provide measurements for an environment are
separate from the measurements themselves. The authorities are more part
of the "predicate" than they are of the object.

The semantic difference is the removal of authorities from
measurement-map in endorsed-triple-record as it appears in
conditional-endorsement-triple-record's consequents.

The authorities are only meant for a matching condition and not as an
authority delegation.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>